requirements.txt
requests==2.18.4
beautifulsoup4==4.6.3
IMDbPY==6.6
tqdm==4.25.0
